added wind direction to  day forecast screen

+++ data/src/main/java/com/lab/esh1n/data/cache/entity/WeatherWithPlace.kt
@@ -29,7 +29,7 @@ class WeatherWithPlace(
 
         var windSpeed: Double,
 
-        var windDegree: Int,
+        var windDegree: Double,
 
         var pressure: Int,
